Featurizer updates
